+++ src/internal/pkg/dbase/model.go
@@ -1,0 +1,6 @@
+package dbase
+
+type StatisticItem struct {
+	Code   string
+	UseCnt int64
+}